+++ test/fixtures/ts/ts3.ts
@@ -1,0 +1,4 @@
+const ts3 = function () {
+  //
+};
+export default ts3;