+++ test/fixtures/ts1.ts
@@ -1,0 +1,3 @@
+import ts2 from './ts2.ts';
+
+ts2();