+++ test/fixtures/ts2.ts
@@ -1,0 +1,10 @@
+// eslint-disable-next-line import/extensions, node/no-missing-import
+import ts from './ts';
+
+ts();
+
+const ts2 = function () {
+  //
+};
+
+export default ts2;