+++ CHANGES.md
@@ -1,5 +1,10 @@
 # CHANGES for `es-file-traverse`
 
+## 0.4.0
+
+- Enhancement: Add `typescript` option for resolving by TypeScript
+    [module resolution algorithm](https://www.typescriptlang.org/docs/handbook/module-resolution.html); currently auto-sets `defaultSourceType` to `module`.
+
 ## 0.3.0
 
 - Breaking change: `babelEslintOptions` -> `parserOptions`